@@ -58,11 +58,11 @@
     "@rollup/plugin-commonjs": "^17.0.0",
     "@rollup/plugin-node-resolve": "^11.0.0",
     "@rollup/plugin-typescript": "^8.2.5",
-    "@storybook/addon-essentials": "^8.1.3",
-    "@storybook/addon-svelte-csf": "^4.1.3",
-    "@storybook/svelte": "^8.1.3",
-    "@storybook/svelte-vite": "^8.1.3",
-    "@sveltejs/vite-plugin-svelte": "^3.1.1",
+    "@storybook/addon-essentials": "^8.6.14",
+    "@storybook/addon-svelte-csf": "^5.0.8",
+    "@storybook/svelte": "8",
+    "@storybook/svelte-vite": "8",
+    "@sveltejs/vite-plugin-svelte": "^6.1.3",
     "@tinymce/beehive-flow": "^0.19.0",
     "@tinymce/eslint-plugin": "^2.3.1",
     "@tsconfig/svelte": "^5.0.4",
@@ -70,7 +70,7 @@
     "@typescript-eslint/parser": "^7.13.0",
     "babel-loader": "^8.2.2",
     "eslint": "^8",
-    "eslint-plugin-svelte": "^2.39.3",
+    "eslint-plugin-svelte": "^3.11.0",
     "gh-pages": "^6.1.1",
     "react": "^18",
     "react-dom": "^18",
@@ -78,17 +78,17 @@
     "rollup-plugin-css-only": "^3.1.0",
     "rollup-plugin-execute": "^1.1.1",
     "rollup-plugin-livereload": "^2.0.0",
-    "rollup-plugin-svelte": "^7.1.5",
+    "rollup-plugin-svelte": "^7.2.3",
     "rollup-plugin-terser": "^7.0.0",
-    "storybook": "^8.1.6",
-    "svelte": "^4.2.17",
-    "svelte-check": "^3.8.0",
-    "svelte-loader": "^3.2.0",
-    "svelte-preprocess": "^5.1.4",
+    "storybook": "8",
+    "svelte": "^5.38.3",
+    "svelte-check": "^4.3.1",
+    "svelte-loader": "^3.2.4",
+    "svelte-preprocess": "^6.0.3",
     "tinymce": "^8.0.0",
     "tslib": "^2.6.2",
-    "typescript": "^5.4.5",
-    "vite": "^5.2.13",
+    "typescript": "^5.9.2",
+    "vite": "^7.1.3",
     "webpack": "^5.76.2"
   },
   "resolutions": {
